install NERDTree, distro dependent package install, tweaks

+++ packages.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+os_release=$(grep -oP "^ID=\K.*" /etc/os-release)
+echo "--> detected ${os_release}"
 
 # ncdu       - ncurses disk usage
 # entr       - file system watcher
@@ -10,13 +12,10 @@
 # shellcheck - linter for shell scripts
 # htop       - nice top
 
-sudo apt install -yqq ncdu entr jq fasd httpie lnav shellcheck htop         
-
-# bat - a nice cat alternative:update
-if [ ! -x "$(command -v bat)" ]; then
-    wget -q -O /tmp/bat.deb https://github.com/sharkdp/bat/releases/download/v0.6.1/bat_0.6.1_amd64.deb
-    sudo dpkg -i /tmp/bat.deb
-    rm -f /tmp/bat.deb
+if [ "${os_release}" == "ubuntu" ]; then
+	sudo apt install -yqq ncdu entr jq fasd httpie lnav shellcheck htop         
+elif [ "${os_release}" == "fedora" ]; then
+	sudo dnf install -y ncdu entr jq fasd httpie lnav ShellCheck htop         
 fi
 
 # NVM
@@ -26,9 +25,13 @@ fi
 
 # ripgrep
 if [ ! -x "$(command -v rg)" ]; then
+  if [ "${os_release}" == "ubuntu" ]; then
     wget -q -O /tmp/ripgrep.deb https://github.com/BurntSushi/ripgrep/releases/download/0.9.0/ripgrep_0.9.0_amd64.deb
     sudo dpkg -i /tmp/ripgrep.deb
     rm -f /tmp/ripgrep.deb
+  elif [ "${os_release}" == "fedora" ]; then
+    sudo dnf install -y ripgrep
+  fi
 fi
 
 cat <<EOF > "${HOME}/.nvm/default-packages"
